refactor(home): extract ToggleButton out of add-tile-drawer.tsx
add-tile-drawer.tsx is a 1100+ line God-component. ToggleButton is a
pure presentational atom (only reads its own props) already imported
cross-file by agent-prompt-list.tsx via the drawer module, so pulling
it into its own file removes an unnecessary coupling to the drawer's
internals without changing any behavior.
